fix(client): Revert recent changes in fetchRelationships 🚑
This commits reverts recent changes:
* Async Association.query
* Taking into account documents returned by fetchRelationships
* Handle same instance of QueryDefinition returned by multiple
Association.query methods

This revert follow a regression which appeared in Banks, and a
new implementation of HasManyTriggers which does not need
async query and returing array of documents anymore.
